Updated basically everything because of some templating in mat4

+++ ftec-testgame/src/math/Delaunay.h
@@ -25,7 +25,7 @@ namespace ftec {
 		const vec2f &getPoint(int index) const { return m_Vertices[index].m_Vertex; }
 		const std::vector<DelaunayVertex<vec2f>> &getVertices() const { return m_Vertices; }
 
-		int getTriangleCount() const { return m_Triangles.size(); }
+		int getTriangleCount() const { return (int) m_Triangles.size(); }
 		const TriangleRef &getTriangleRef(int index) const { return m_Triangles[index]; }
 		triangle2f getTriangle(int index) const;
